chore: cleanup Mattermost webhook handler
